cleanup(ε.3): extract decode_block/coarse_sync.rs (Costas correlation search)
Third slice of the `decode_block.rs` per-stage split — the biggest
single piece. Moves the full coarse-sync stage into a sibling
submodule:

- `coarse_sync()`, `coarse_sync_with_allsum()` — public entry points
- `coarse_allsum_len()`, `precompute_coarse_allsum()`,
  `precompute_coarse_allsum_into()` — public allsum-buffer helpers
  consumed by `embedded-shared::stage1_inc` and the m5stack apps
- `fill_coarse_allsum()` — private allsum-fill implementation
  shared between the public precompute helpers and `coarse_sync_inner`
- `coarse_sync_inner()` — the ~330-line Costas correlation + score
  loop + dedupe (peak picker, parabolic interpolation, `MLAG`
  greedy NMS, 4 Hz / 40 ms greedy dedupe)

External-caller path preserved: parent re-exports the 5 public
entries via `pub use coarse_sync::*`. The private `fill_coarse_allsum`
and `coarse_sync_inner` are not exposed.

`NSSY` / `ratio_eps` / `sync_lag_s` imports dropped from the parent's
`use types::*` block — all three are now reached only from inside
coarse_sync.rs.

Acceptance:
- `cargo check` clean on all 4 feature cells (default, embedded ship,
  host fixed-point bench, embedded `nstep-half`)
- FT8 qso3_busy golden tests pass: WSJT-X AP-off floor, JTDX recall
  floor, AP-on strict-superset, AP-on subtract diag
- 7 `ft8::decode_block::tests::*` lib tests pass
- `cargo clippy --features full --lib` clean, `cargo fmt --check` clean
- Diff: decode_block.rs −520 / +8, new decode_block/coarse_sync.rs +537
  (verbatim copy + cross-module imports the moved code needs)

`decode_block.rs` is now down to 2,516 lines (started at 3,517).
Remaining ε slices: fill_symbol_spectra (ε.4), process_candidates
(ε.5), osd_strategy (ε.6).

Stacks on #78 (ε.2). Rebases to main when #78 lands.
